fix: Fixed VH button and add new member button positon
